requirements.txt
Django==2.0.2
djangorestframework==3.7.7
psycopg2==2.7.4